temp move back import-shared-modification to import-composite
